Refactor tests, add generic audit demo
